backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
pymilvus==2.4.0
pydantic==2.5.0
pydantic-settings==2.1.0
PyPDF2==3.0.1
python-docx==0.8.11
minio==7.1.16
python-multipart==0.0.6
PyJWT==2.8.0
requests==2.31.0
marshmallow<4
celery==5.3.4
redis==5.0.1
magic-pdf==0.7.1
pdfplumber>=0.11.0
